% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABMv4.R
\name{covid_abm_v3}
\alias{covid_abm_v3}
\title{ABM V3 with testing}
\usage{
covid_abm_v3(
  bta_fx,
  bta_hh,
  bta_scl,
  bta_work,
  bta_other,
  E0,
  Ip0,
  Ia0,
  Im0,
  Imh0,
  Ih0,
  R0,
  D0,
  agents_dt,
  cbg_cdf_list,
  cbg_ids,
  stay_home_dt,
  visitors,
  visitors_list,
  visitor_mult_testing,
  visitor_mult_sfgrph,
  t0,
  t.tot,
  dt,
  day_of_week_fx,
  time_of_day_fx,
  SiP.start,
  scl.close,
  part.reopen,
  mask.start,
  mask.red,
  mask_fx,
  social_fx,
  q_prob_contact,
  q_prob_resinf,
  q_prob_symptoms,
  q_prob_testpos,
  q_prob_adapt,
  q_dur_fx,
  known_contact_prob,
  other_ill,
  init_other,
  other_bta,
  other_symp_dur_fx,
  testing,
  test_start,
  test_delay_fx,
  tests_pp_fx,
  tests_pub,
  tests_wknd,
  test_probs_pub_fx,
  test_probs_pvt_fx,
  symp_mult,
  cont_mult,
  res_mult,
  nosymp_state_mult,
  symp_state_mult,
  hosp_mult,
  test.red,
  adaptive,
  adapt_start,
  adapt_freq,
  adapt_site_fx,
  adapt_site_geo,
  n_adapt_sites,
  adapt_site_test_criteria,
  adapt_site_mult,
  adapt_site_delay_fx,
  verbose,
  store_extra
)
}
\arguments{
\item{bta_fx}{function that takes time step as input returns beta at that time step}

\item{bta_hh}{houshold transmission probability multiplier}

\item{bta_scl}{school transmission probability multiplier}

\item{bta_work}{work transmission probability multiplier}

\item{bta_other}{other transmission probability multiplier}

\item{E0}{number of initial exposed agents}

\item{Ip0}{number of initial pre-symptomatic infections agents}

\item{Ia0}{number of initial asymptomatic infectious agents}

\item{Im0}{number of initial mildy symptomatic agents}

\item{Imh0}{number of initial mildly symptomatic, will become severely symptomatic agents}

\item{Ih0}{number of initial hospitalized agents}

\item{R0}{number of initial recovered agents}

\item{D0}{number of initial diceased agents}

\item{agents_dt}{data.table of agents with necessary characteristics including: TODO: FILL THESE IN}

\item{cbg_cdf_list}{list of matrices of dim cbgs x cbgs in which entries represent probabilities of moving from CBG i to CBG j. Used in `GetCBGVisit`/`GetCBGVisit_cpp` function}

\item{cbg_ids}{vector relating row numbers of `cbg_cdf_list` to actual cbg codes}

\item{stay_home_dt}{data table with columns `Date`, `origin_census_block_group`, and `pct_home` (E.g. derived from safegraph data) to use for social distancing/stay at home compliance}

\item{visitors}{logical of whether to model outside visitors}

\item{visitors_list}{list of data frames/tables with columns corresponding to census block group, number of external visitors to that cbg, county from which they visited, and population and infection characteristics of that county. column inf_prob represents incidence in that county (new cases identified per population) and is used as probability a visiting agent is infectious along with `visitor_mult_testing` to correct for ratio of true incidence to incidence from testing data and `visitor_mult_sfgrph` to correct for ratio of safegraph devices tracked to total population}

\item{visitor_mult_testing}{multiplier to adjust probability visiting agent is infected based on ratio of true incidence to incidence identified via testing}

\item{visitor_mult_sfgrph}{multiplier to adjust total number of visitors from safegraph number of visitors based on ratio of tracked safegraph devices to total movement of people. THis is quite subjective as we don't really know much about Safegraph's panel and how it realtes to underlying population characteristics}

\item{t0}{start date of simulation}

\item{t.tot}{total time to run simulation (in days)}

\item{dt}{time step (defaults to 1/6, needs editing if otherwise)}

\item{day_of_week_fx}{function which takes t returns character of day of week (U,M,T,W,R,F,S)}

\item{SiP.start}{Date that shelter in place started}

\item{scl.close}{date that schools were closed}

\item{part.reopen}{Date that partial reopen allows essential workers back to work}

\item{mask.start}{Date that mask mandate was introduced}

\item{mask.red}{Reduction in transmission probability from wearing mask}

\item{mask_fx}{fnction to assign probabilities of wearing a mask following mask mandate}

\item{social_fx}{function to generate random sociality metrics}

\item{q_prob_contact}{probability an agent will quarantine because of known contact with infectious agent}

\item{q_prob_resinf}{probability an agent will quarantine because of shared residence with infectious agent}

\item{q_prob_symptoms}{probability an agent will quarantine because of symptoms on a per day basis: e.g. one full day of symptoms = q_prob_symptoms probability of isolating. Note that agents will consider this 4 times per day, so probability should adjust for multiple opportunities}

\item{q_prob_testpos}{probability an agent will quarantine because of positive test result}

\item{q_prob_adapt}{probability multiplier agent will quarantine if "exposed" to an adaptive testing site}

\item{q_dur_fx}{function to generate random quarantine durations}

\item{known_contact_prob}{proportional increase in probability that a contact with an infectious individual is known. 0 implies known contact occurs at same rate as FOI, implying very low probability that an infectious contact will be known 1 implies probability is double FOI. Since FOIs will be quite small, larger multipliers are suggested}

\item{other_ill}{logical of whether to include circulation of other illness that causes COVID-like symptoms}

\item{init_other}{initial number of individuals with other pathogens that cause COVID-like illness}

\item{other_bta}{Relative transmission rate of "other" pathogens that cause covid-like symptoms and are transmitted across the same networks}

\item{other_symp_dur_fx}{function to generate duration of symptoms for non-COVID illness}

\item{testing}{TRUE/FALSE for whether to conduct testing}

\item{test_start}{Date at which testing begins}

\item{test_delay_fx}{function to generate time delay from test to notification}

\item{tests_pp_fx}{function returning the number of tests available per agent on day t since test start}

\item{tests_pub}{proportion of tests that are public vs private}

\item{tests_wknd}{percent of weekday tests available on weekends}

\item{test_probs_pub_fx}{function which returns testing probability for individuals to get tested at public site}

\item{test_probs_pvt_fx}{function which returns testing probability for individuals to get tested at prvate site}

\item{symp_mult}{multiplier for test probability for agents with symptoms. Acts on a per-day with symptoms basis, so 2 days with symptoms * `symp_mult=10` would lead to 20x increase in test prob}

\item{cont_mult}{multiplier for test probability for agents with suspeceted contact}

\item{res_mult}{multiplier for test probability for agents with known infection in their residence}

\item{nosymp_state_mult}{multiplier for test probability for agents who are infected but not shopwing symptoms (Ia or Ip)}

\item{symp_state_mult}{multiplier for test probability for agents who are infected andshowing symptoms (Im or Imh)}

\item{hosp_mult}{multiplier for test probability for agents who are infected and hospitalized}

\item{test.red}{reduction in transmission probability if agents has tested positive}

\item{adaptive}{TRUE/FALSE for whether adaptive testing should be implemented with `adapt_site_fx` in simulation}

\item{adapt_start}{time point at which to start adaptive testing}

\item{adapt_freq}{frequency with which to check test reports and generate new test site}

\item{adapt_site_fx}{function to determine when and where to place adaptive testing sites}

\item{adapt_site_geo}{geography to base adaptive sites on. Can be `cbg`, `ct`, `nbhd`, or `zip`}

\item{n_adapt_sites}{how many adaptive sites available to place}

\item{adapt_site_test_criteria}{character either `n_pos`or `pct_pos`for whteher targeted sites should be added based on census tract with highest number positive tests or highest percent positive tests}

\item{adapt_site_mult}{increase in testing probability if test site is placed in residence ct}

\item{adapt_site_delay_fx}{function to generate time delay for adaptive sites}

\item{verbose}{should detailed info at each time step be printed?}

\item{store_extra}{should extra metrics including number of other illnesses, % staying home, and % isolating be stored and returned? Good for debugging}
}
\value{
list with two objects: an epi curve with counts of agents in each state through time and a dataframe with infection reports for every new infection (e.g. entry created whenever agent transitions from S to E)
}
\description{
Function to run agent based model
}
