% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infection_Functions.R
\name{foi_infect}
\alias{foi_infect}
\title{Simulate Infection}
\usage{
foi_infect(foi)
}
\arguments{
\item{foi}{Person's FOI}
}
\value{
binary of infection occurring or not
}
\description{
Randomly determine if infection occurs from FOI
}
