% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Mvmt_Functions.R
\name{GetCTVisit}
\alias{GetCTVisit}
\title{Community location function}
\usage{
GetCTVisit(mat_cdf, cts, agent_cts)
}
\arguments{
\item{mat_cdf}{cumulative probability matrix with rows corresponding to CDF of individual residing in row and visiting column}

\item{cbgs}{cbg names coresponding to rows of cbg_cdf}

\item{agent_cbgs}{cbgs of agents which are candidates to interact in the community}
}
\value{

}
\description{
Draws random uniform variable between 0 and 1, looks in CDF matrix to determine corresponding probability of visiting CBG, finds corresponding CBG visited in index matrix. See also `GetCBGVisit_cpp` for faster implementation
}
