% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABMv4.R
\name{mask_fx}
\alias{mask_fx}
\title{Get mask compliance probabilities}
\usage{
mask_fx(n_agents)
}
\arguments{
\item{n_agents}{number of agents to generate probabilities for}
}
\value{
vector of length n_agents with randomly generated probabilities of wearing mask following implementation of mask mandate
}
\description{
Function to generate random mask compliance
}
