% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABMv4.R
\name{other_location}
\alias{other_location}
\title{Simulate location of individuals who are not workers or school-aged}
\usage{
other_location(
  inf.state,
  SiP,
  home_prob,
  age,
  sociality,
  res_type,
  res_id,
  comm_id
)
}
\arguments{
\item{inf.state}{Infection state, one of S, E, Ip, Ia, Im, Imh, Ih, R}

\item{SiP}{shelter in place active? 1/0}

\item{age}{age of person}

\item{sociality}{agent sociality}

\item{res_type}{type of residence (H, C, P, or N)}

\item{res_id}{id of this individual's residence}

\item{comm_id}{id of this individual's community}

\item{time_day}{time of day (night or day)}
}
\value{
location of this individual in the time step with options being the id corresponding to their home (H), work (W), school (S), or community (C)
}
\description{
Simulate location of non workers or school children
}
