% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABMv4.R
\name{other_symp_dur_fx}
\alias{other_symp_dur_fx}
\title{Duration of COVID-like symptoms caused by non-SARS-CoV2 infection}
\usage{
other_symp_dur_fx(n_agents)
}
\arguments{
\item{n_agents}{number of agents to draw for}
}
\value{
vector of length `n_agents`
}
\description{
Function to generate duration of COVID-like symptoms caused by non-SARS-CoV2 infection
}
