% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infection_Functions.R
\name{p_symp}
\alias{p_symp}
\title{Probability of symptomatic infection}
\usage{
p_symp(age)
}
\arguments{
\item{age}{age of person}
}
\value{
numeric of probability of clinical (symptomatic).
}
\description{
Age-dependent probabilities of having clinical (symptomatic) infection. Used in `next.state`. From combination of UeS data for 0-49 and best guesstimate/input from COVASIM for older age groups
}
