% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABMv4.R
\name{q_dur_fx}
\alias{q_dur_fx}
\title{Quarantine duration}
\usage{
q_dur_fx(n_agents)
}
\arguments{
\item{n_agents}{number of agents to generate probabilities for}
}
\value{
vector of length n_agents with randomly generated duration of quarantine following symptom onset/contact/positive test
}
\description{
Function to generate random number of days of quarantine
}
