% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{q_prob_fx}
\alias{q_prob_fx}
\title{q_prob_fx}
\usage{
q_prob_fx(
  contact,
  t_since_contact,
  q_prob_contact,
  res_inf,
  q_prob_resinf,
  t_symptoms,
  q_prob_symptoms,
  tested,
  infector,
  q_prob_testpos,
  essential,
  q_prob_essential
)
}
\value{
Vector of quarantine probabilities for all agents to make quarantine decisions
}
\description{
q_prob_fx
}
