% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Mvmt_Functions.R
\name{scl_wrk_loc}
\alias{scl_wrk_loc}
\title{Sample location}
\usage{
scl_wrk_loc(l_res, l_com, l_scl_wrk, p_res, p_com)
}
\arguments{
\item{l_res}{residence location id}

\item{l_com}{community location id}

\item{p_res}{probability of being at residence}

\item{p_com}{probability of being in community}

\item{l_scl}{work/school location id}
}
\value{
vector of location ids
}
\description{
Function to determine agent's location given present model conditions. Works for agents with 3 possible locations residence/home, community and either work or school. Function passes location ids plus probability of home and community, with probability work/ school then implied
}
