% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABMv4.R
\name{social_fx}
\alias{social_fx}
\title{Sociality}
\usage{
social_fx(n_agents)
}
\arguments{
\item{n_agents}{number of agents to generate probabilities for}
}
\value{
vector of length n_agents with randomly generated sociality
}
\description{
Function to generate random sociality metric
}
