% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABMv4.R
\name{test_delay_fx}
\alias{test_delay_fx}
\title{Delay between getting tested and getting results}
\usage{
test_delay_fx(n_agents)
}
\arguments{
\item{n_agents}{number of agents to generate test delays for}
}
\value{
vector of length n_agents with randomly generated time delay between test and result
}
\description{
Function to generate random number of days from tested to test result
}
