% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABMv4.R
\name{test_probs_pvt_fx}
\alias{test_probs_pvt_fx}
\title{Test Probability function for private tests}
\usage{
test_probs_pvt_fx(
  income_bracket,
  race,
  essential,
  t_symptoms,
  state,
  t_since_contact,
  res_inf,
  res_type,
  tests_avail,
  symp_mult,
  cont_mult,
  res_mult,
  nosymp_state_mult,
  symp_state_mult,
  hosp_mult
)
}
\arguments{
\item{income_bracket}{vector of agent income brackets (in quartiles)}

\item{race}{vector of agent races}

\item{essential}{binary of whether agent is an essential worker}

\item{t_symptoms}{vector of time experiencing symptoms}

\item{t_since_contact}{vector of time left that person with known contact will consider contact a factor in testing decision}

\item{res_inf}{Number of identified infectious individuals in residence}

\item{res_type}{vector of residence types, "H", "N", "P"}

\item{tests_avail}{number of tests available - creates multiplier to serve as proxy for how good identification of infected individuals is}

\item{symp_mult}{multiplier for test probability for agents with symptoms. Acts on a per-day with symptoms basis, so 2 days with symptoms * `symp_mult=10` would lead to 20x increase in test prob}

\item{cont_mult}{multiplier for test probability for agents with suspeceted contact}

\item{res_mult}{multiplier for test probability for agents with known infection in their residence}

\item{nosymp_state_mult}{multiplier for test probability for agents who are infected but not shopwing symptoms (Ia or Ip)}

\item{symp_state_mult}{multiplier for test probability for agents who are infected andshowing symptoms (Im or Imh)}

\item{hosp_mult}{multiplier for test probability for agents who are infected and hospitalized}
}
\value{
vector of length n_agents with generated test probabilities
}
\description{
Function to generate probability of getting tested based on income, race, symptoms, and known contacts
}
