% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{unpack_list}
\alias{unpack_list}
\title{Unpack elements of list into global environment}
\usage{
unpack_list(l)
}
\arguments{
\item{l}{list (or list of lists) to unpack}
}
\value{
assigns elements of list(s) to global environment
}
\description{
Taken from https://stackoverflow.com/questions/26168592/recursively-send-list-variables-to-the-global-environment and turned quiet
}
