% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Mvmt_Functions.R
\name{visitors_to_agents}
\alias{visitors_to_agents}
\title{Convert safegraph visitor panel to agents}
\usage{
visitors_to_agents(visits_dt, visitor_mult_testing, visitor_mult_sfgrph)
}
\arguments{
\item{visits_dt}{data.table of visitors to census tracts/block groups}

\item{visitor_mult_testing}{multiplier on true infection rate of visitorsto account for fact that not all infections are tested}

\item{visitor_mult_sfgrph}{multiplier on true number of visitors to account for fact that not all people are captured in safegraph paenel data}
}
\value{
dataframe of visiting agents to append to regular agents
}
\description{
takes number of visitors from county to ct and converts to a data table of agents to merge with resident agents. Assumes agents spend time in community and are either infected or not. All other characteristics left NA
}
