% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Mvmt_Functions.R
\name{worker_location}
\alias{worker_location}
\title{Simulate worker locations}
\usage{
worker_location(
  inf.state,
  SiP,
  home_prob,
  time_day,
  day_week,
  age,
  essential,
  sociality,
  res_id,
  work_id,
  comm_id
)
}
\arguments{
\item{inf.state}{Infection state, one of S, E, Ip, Ia, Im, Imh, Ih, R}

\item{SiP}{shelter in place active? 1/0}

\item{home_prob}{probability agent in ct is at home from safegraph metrics}

\item{time_day}{time of day (night or day)}

\item{day_week}{day of the week (U, M, T, W, R, F, or S)}

\item{age}{age of person}

\item{essential}{is the individual in an essential workforce? 1/0}

\item{sociality}{agent sociality}

\item{res_id}{id of this individual's residence}

\item{work_id}{id of this individual's workplace}

\item{comm_id}{id of this individual's community}
}
\value{
location of this individual in the time step with options being the id corresponding to their home (H), work (W), or community (C)
}
\description{
Simulate location of school-aged children depending on agent characteristics, NPIs, and time of day
}
