% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Mvmt_Functions.R
\name{other_location}
\alias{other_location}
\title{Simulate location of individuals who are not workers or school-aged}
\usage{
other_location(inf.state, SiP, home_prob, age, sociality, hhid, ct)
}
\arguments{
\item{inf.state}{Infection state, one of S, E, Ip, Ia, Im, Imh, Ih, R}

\item{SiP}{shelter in place active? 1/0}

\item{age}{age of person}

\item{sociality}{agent sociality}

\item{hhid}{id of this individual's residence}

\item{time_day}{time of day (night or day)}

\item{comm_id}{id of this individual's community}
}
\value{
location of this individual in the time step with options being the id corresponding to their home (H), work (W), school (S), or community (C)
}
\description{
Simulate location of non workers or school children
}
