% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_probs_fx}
\alias{test_probs_fx}
\title{test_probs_fx}
\usage{
test_probs_fx(
  income,
  income_mult,
  hpi_hc,
  hpi_mult,
  essential,
  essential_prob,
  t_symptoms,
  symp_mult,
  state,
  nosymp_state_mult,
  symp_state_mult,
  t_since_contact,
  cont_mult,
  res_inf,
  res_mult,
  adapt_site,
  adapt_site_mult,
  tests_avail,
  case_find_mult,
  hosp_mult
)
}
\value{
Vector of test probabilities for all agents to weight sampling to determine who is tested
}
\description{
test_probs_fx
}
