% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABMv4.R
\name{visitors_to_agents}
\alias{visitors_to_agents}
\title{Convert safegraph visitor panel to agents}
\usage{
visitors_to_agents(visits_dt, visitor_mult_testing, visitor_mult_sfgrph)
}
\arguments{
\item{visits_dt}{cumulative probability matrix with rows corresponding to CDF of individual residing in row and visiting column}

\item{visitor_mult_testing}{cbg names coresponding to rows of cbg_cdf}

\item{visitor_mult_sfgrph}{cbgs of agents which are candidates to interact in the community}
}
\value{

}
\description{
takes number of visitors from county to cbg and converts to a data table of agents to merge with resident agents. Assumes agents spend time in community and are either infected or not. All other characteristics left NA
}
